Breaking: RFC5424 signature changes
This commit introduces a refactoring of RFC5424 machines in order to
respect newly introduced interfaces.

The following are the new method signatures:

* rfc5424.NewMachine(opts ...syslog.MachineOptions) syslog.Machine
* rfc5424.NewParser(opts ...syslog.MachineOptions) syslog.Machine
* Parse(input[]byte) (syslog.Message, error)
* WithBestEffort() syslog.MachineOption

For both there is no more need to pass a pointer to signify the best effort mode
on or off. An option has been introduced for such reason.

Signed-off-by: Leonardo Di Donato <leodidonato@gmail.com>